Add Public.com trading + free market data, zero-subscription data stack
- Public.com broker plugin upgraded from documented scaffold to full
  plugin against the official Trading API: token auth, portfolio v2,
  single-leg and multi-leg orders (UUID idempotency, GTC->GTD mapping,
  extended hours), cancel/status/open orders. Refuses paper:true since
  Public has no paper environment.
- New public_data market data provider: real-time quotes, OHLCV bars,
  and option chains with greeks, free with a Public account; registered
  in BUILTIN_PROVIDERS with crypto symbol support.
- Shared _post_json helper on MarketDataProvider; JSON vault credentials
  now accept a 'secret' key field so one entry serves broker + data.
- Example config and docs reworked around an all-free data stack
  (Public/Finnhub/TwelveData/AlphaVantage first; Polygon optional paid).
- mypy --strict now fully clean (fixed 20 latent typing errors across
  strategies, backtest engine, scheduler, Schwab plugin, CLI, kernel).
- 20 new unit tests for the Public broker payloads/parsing and data
  provider (121 total).

+++ src/aegis_trader/strategy/base.py
@@ -76,4 +76,4 @@ def realized_vol(closes: list[float], window: int = 20) -> float | None:
         return None
     mean = sum(rets) / len(rets)
     var = sum((r - mean) ** 2 for r in rets) / len(rets)
-    return (var ** 0.5) * (252 ** 0.5)
+    return float((var ** 0.5) * (252 ** 0.5))
